Add icon names as comments; add cancel icon
